DREAM/requirements.txt
Keras==2.3.1
keras-tuner==1.0.2
tensorflow-gpu==2.4.3
matplotlib==3.3.0
tensorflow_datasets==2.1.0
autokeras==1.0.12
nni==3.0